pkg/proc: File organization
