+++ pkg/proc/threads.go
@@ -5,12 +5,9 @@ import (
 	"fmt"
 	"go/ast"
 	"go/token"
-	"path/filepath"
 	"reflect"
-	"strings"
 
 	"github.com/go-delve/delve/pkg/dwarf/godwarf"
-	"github.com/go-delve/delve/pkg/dwarf/reader"
 )
 
 // Thread represents a thread.
@@ -113,368 +110,6 @@ type ErrNoSourceForPC struct {
 
 func (err *ErrNoSourceForPC) Error() string {
 	return fmt.Sprintf("no source for PC %#x", err.pc)
-}
-
-// Set breakpoints at every line, and the return address. Also look for
-// a deferred function and set a breakpoint there too.
-// If stepInto is true it will also set breakpoints inside all
-// functions called on the current source line, for non-absolute CALLs
-// a breakpoint of kind StepBreakpoint is set on the CALL instruction,
-// Continue will take care of setting a breakpoint to the destination
-// once the CALL is reached.
-//
-// Regardless of stepInto the following breakpoints will be set:
-// - a breakpoint on the first deferred function with NextDeferBreakpoint
-//   kind, the list of all the addresses to deferreturn calls in this function
-//   and condition checking that we remain on the same goroutine
-// - a breakpoint on each line of the function, with a condition checking
-//   that we stay on the same stack frame and goroutine.
-// - a breakpoint on the return address of the function, with a condition
-//   checking that we move to the previous stack frame and stay on the same
-//   goroutine.
-//
-// The breakpoint on the return address is *not* set if the current frame is
-// an inlined call. For inlined calls topframe.Current.Fn is the function
-// where the inlining happened and the second set of breakpoints will also
-// cover the "return address".
-//
-// If inlinedStepOut is true this function implements the StepOut operation
-// for an inlined function call. Everything works the same as normal except
-// when removing instructions belonging to inlined calls we also remove all
-// instructions belonging to the current inlined call.
-func next(dbp *Target, stepInto, inlinedStepOut bool) error {
-	backward := dbp.GetDirection() == Backward
-	selg := dbp.SelectedGoroutine()
-	curthread := dbp.CurrentThread()
-	topframe, retframe, err := topframe(selg, curthread)
-	if err != nil {
-		return err
-	}
-
-	if topframe.Current.Fn == nil {
-		return &ErrNoSourceForPC{topframe.Current.PC}
-	}
-
-	if backward && retframe.Current.Fn == nil {
-		return &ErrNoSourceForPC{retframe.Current.PC}
-	}
-
-	// sanity check
-	if inlinedStepOut && !topframe.Inlined {
-		panic("next called with inlinedStepOut but topframe was not inlined")
-	}
-
-	success := false
-	defer func() {
-		if !success {
-			dbp.ClearInternalBreakpoints()
-		}
-	}()
-
-	ext := filepath.Ext(topframe.Current.File)
-	csource := ext != ".go" && ext != ".s"
-	var thread MemoryReadWriter = curthread
-	var regs Registers
-	if selg != nil && selg.Thread != nil {
-		thread = selg.Thread
-		regs, err = selg.Thread.Registers(false)
-		if err != nil {
-			return err
-		}
-	}
-
-	sameGCond := SameGoroutineCondition(selg)
-
-	var firstPCAfterPrologue uint64
-
-	if backward {
-		firstPCAfterPrologue, err = FirstPCAfterPrologue(dbp, topframe.Current.Fn, false)
-		if err != nil {
-			return err
-		}
-		if firstPCAfterPrologue == topframe.Current.PC {
-			// We don't want to step into the prologue so we just execute a reverse step out instead
-			if err := stepOutReverse(dbp, topframe, retframe, sameGCond); err != nil {
-				return err
-			}
-
-			success = true
-			return nil
-		}
-
-		topframe.Ret, err = findCallInstrForRet(dbp, thread, topframe.Ret, retframe.Current.Fn)
-		if err != nil {
-			return err
-		}
-	}
-
-	text, err := disassemble(thread, regs, dbp.Breakpoints(), dbp.BinInfo(), topframe.Current.Fn.Entry, topframe.Current.Fn.End, false)
-	if err != nil && stepInto {
-		return err
-	}
-
-	retFrameCond := andFrameoffCondition(sameGCond, retframe.FrameOffset())
-	sameFrameCond := andFrameoffCondition(sameGCond, topframe.FrameOffset())
-	var sameOrRetFrameCond ast.Expr
-	if sameGCond != nil {
-		if topframe.Inlined {
-			sameOrRetFrameCond = sameFrameCond
-		} else {
-			sameOrRetFrameCond = &ast.BinaryExpr{
-				Op: token.LAND,
-				X:  sameGCond,
-				Y: &ast.BinaryExpr{
-					Op: token.LOR,
-					X:  frameoffCondition(topframe.FrameOffset()),
-					Y:  frameoffCondition(retframe.FrameOffset()),
-				},
-			}
-		}
-	}
-
-	if stepInto && !backward {
-		err := setStepIntoBreakpoints(dbp, text, topframe, sameGCond)
-		if err != nil {
-			return err
-		}
-	}
-
-	if !backward {
-		_, err = setDeferBreakpoint(dbp, text, topframe, sameGCond, stepInto)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Add breakpoints on all the lines in the current function
-	pcs, err := topframe.Current.Fn.cu.lineInfo.AllPCsBetween(topframe.Current.Fn.Entry, topframe.Current.Fn.End-1, topframe.Current.File, topframe.Current.Line)
-	if err != nil {
-		return err
-	}
-
-	if backward {
-		// Ensure that pcs contains firstPCAfterPrologue when reverse stepping.
-		found := false
-		for _, pc := range pcs {
-			if pc == firstPCAfterPrologue {
-				found = true
-				break
-			}
-		}
-		if !found {
-			pcs = append(pcs, firstPCAfterPrologue)
-		}
-	}
-
-	if !stepInto {
-		// Removing any PC range belonging to an inlined call
-		frame := topframe
-		if inlinedStepOut {
-			frame = retframe
-		}
-		pcs, err = removeInlinedCalls(dbp, pcs, frame)
-		if err != nil {
-			return err
-		}
-	}
-
-	if !csource {
-		var covered bool
-		for i := range pcs {
-			if topframe.Current.Fn.Entry <= pcs[i] && pcs[i] < topframe.Current.Fn.End {
-				covered = true
-				break
-			}
-		}
-
-		if !covered {
-			fn := dbp.BinInfo().PCToFunc(topframe.Ret)
-			if selg != nil && fn != nil && fn.Name == "runtime.goexit" {
-				return nil
-			}
-		}
-	}
-
-	for _, pc := range pcs {
-		if _, err := allowDuplicateBreakpoint(dbp.SetBreakpoint(pc, NextBreakpoint, sameFrameCond)); err != nil {
-			dbp.ClearInternalBreakpoints()
-			return err
-		}
-	}
-
-	if stepInto && backward {
-		err := setStepIntoBreakpointsReverse(dbp, text, topframe, sameGCond)
-		if err != nil {
-			return err
-		}
-	}
-
-	if !topframe.Inlined {
-		// Add a breakpoint on the return address for the current frame.
-		// For inlined functions there is no need to do this, the set of PCs
-		// returned by the AllPCsBetween call above already cover all instructions
-		// of the containing function.
-		bp, err := dbp.SetBreakpoint(topframe.Ret, NextBreakpoint, retFrameCond)
-		if err != nil {
-			if _, isexists := err.(BreakpointExistsError); isexists {
-				if bp.Kind == NextBreakpoint {
-					// If the return address shares the same address with one of the lines
-					// of the function (because we are stepping through a recursive
-					// function) then the corresponding breakpoint should be active both on
-					// this frame and on the return frame.
-					bp.Cond = sameOrRetFrameCond
-				}
-			}
-			// Return address could be wrong, if we are unable to set a breakpoint
-			// there it's ok.
-		}
-		if bp != nil {
-			configureReturnBreakpoint(dbp.BinInfo(), bp, &topframe, retFrameCond)
-		}
-	}
-
-	if bp := curthread.Breakpoint(); bp.Breakpoint == nil {
-		curthread.SetCurrentBreakpoint(false)
-	}
-	success = true
-	return nil
-}
-
-func setStepIntoBreakpoints(dbp Process, text []AsmInstruction, topframe Stackframe, sameGCond ast.Expr) error {
-	for _, instr := range text {
-		if instr.Loc.File != topframe.Current.File || instr.Loc.Line != topframe.Current.Line || !instr.IsCall() {
-			continue
-		}
-
-		if instr.DestLoc != nil {
-			if err := setStepIntoBreakpoint(dbp, []AsmInstruction{instr}, sameGCond); err != nil {
-				return err
-			}
-		} else {
-			// Non-absolute call instruction, set a StepBreakpoint here
-			if _, err := allowDuplicateBreakpoint(dbp.SetBreakpoint(instr.Loc.PC, StepBreakpoint, sameGCond)); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func setStepIntoBreakpointsReverse(dbp Process, text []AsmInstruction, topframe Stackframe, sameGCond ast.Expr) error {
-	// Set a breakpoint after every CALL instruction
-	for i, instr := range text {
-		if instr.Loc.File != topframe.Current.File || !instr.IsCall() || instr.DestLoc == nil || instr.DestLoc.Fn == nil {
-			continue
-		}
-
-		if fn := instr.DestLoc.Fn; strings.HasPrefix(fn.Name, "runtime.") && !isExportedRuntime(fn.Name) {
-			continue
-		}
-
-		if nextIdx := i + 1; nextIdx < len(text) {
-			if _, err := allowDuplicateBreakpoint(dbp.SetBreakpoint(text[nextIdx].Loc.PC, StepBreakpoint, sameGCond)); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func FindDeferReturnCalls(text []AsmInstruction) []uint64 {
-	const deferreturn = "runtime.deferreturn"
-	deferreturns := []uint64{}
-
-	// Find all runtime.deferreturn locations in the function
-	// See documentation of Breakpoint.DeferCond for why this is necessary
-	for _, instr := range text {
-		if instr.IsCall() && instr.DestLoc != nil && instr.DestLoc.Fn != nil && instr.DestLoc.Fn.Name == deferreturn {
-			deferreturns = append(deferreturns, instr.Loc.PC)
-		}
-	}
-	return deferreturns
-}
-
-// Removes instructions belonging to inlined calls of topframe from pcs.
-// If includeCurrentFn is true it will also remove all instructions
-// belonging to the current function.
-func removeInlinedCalls(dbp Process, pcs []uint64, topframe Stackframe) ([]uint64, error) {
-	image := topframe.Call.Fn.cu.image
-	dwarf := image.dwarf
-	irdr := reader.InlineStack(dwarf, topframe.Call.Fn.offset, 0)
-	for irdr.Next() {
-		e := irdr.Entry()
-		if e.Offset == topframe.Call.Fn.offset {
-			continue
-		}
-		ranges, err := dwarf.Ranges(e)
-		if err != nil {
-			return pcs, err
-		}
-		for _, rng := range ranges {
-			pcs = removePCsBetween(pcs, rng[0], rng[1], image.StaticBase)
-		}
-		irdr.SkipChildren()
-	}
-	return pcs, irdr.Err()
-}
-
-func removePCsBetween(pcs []uint64, start, end, staticBase uint64) []uint64 {
-	out := pcs[:0]
-	for _, pc := range pcs {
-		if pc < start+staticBase || pc >= end+staticBase {
-			out = append(out, pc)
-		}
-	}
-	return out
-}
-
-func setStepIntoBreakpoint(dbp Process, text []AsmInstruction, cond ast.Expr) error {
-	if len(text) <= 0 {
-		return nil
-	}
-
-	instr := text[0]
-
-	if instr.DestLoc == nil {
-		// Call destination couldn't be resolved because this was not the
-		// current instruction, therefore the step-into breakpoint can not be set.
-		return nil
-	}
-
-	fn := instr.DestLoc.Fn
-
-	// Skip unexported runtime functions
-	if fn != nil && strings.HasPrefix(fn.Name, "runtime.") && !isExportedRuntime(fn.Name) {
-		return nil
-	}
-
-	//TODO(aarzilli): if we want to let users hide functions
-	// or entire packages from being stepped into with 'step'
-	// those extra checks should be done here.
-
-	pc := instr.DestLoc.PC
-
-	// Skip InhibitStepInto functions for different arch.
-	if dbp.BinInfo().Arch.InhibitStepInto(dbp.BinInfo(), pc) {
-		return nil
-	}
-
-	// We want to skip the function prologue but we should only do it if the
-	// destination address of the CALL instruction is the entry point of the
-	// function.
-	// Calls to runtime.duffzero and duffcopy inserted by the compiler can
-	// sometimes point inside the body of those functions, well after the
-	// prologue.
-	if fn != nil && fn.Entry == instr.DestLoc.PC {
-		pc, _ = FirstPCAfterPrologue(dbp, fn, false)
-	}
-
-	// Set a breakpoint after the function's prologue
-	if _, err := allowDuplicateBreakpoint(dbp.SetBreakpoint(pc, NextBreakpoint, cond)); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func getGVariable(thread Thread) (*Variable, error) {
